bug 1316981, more extensive tests
This uses the traverse code in ast.Node to find all nodes
with _pos, and checks if they're matching what's in the node.

For Placeable and Entity, that's not in the AST, so the code
adds them to passed-in arrays, which can then be checked manually.

The comment check is a bit unspecific about whitespace between
the # and the node.content. Not sure if the ast should or shouldn't
normalize space.

The Section check fails. Other entries contain their trailing
newline, and looking at the spec, I think that sections should, too.
They don't right now, though.